+++ pluginMeasureCpu.py
