+++ pluginMeasurePower.py
